--- src/main/java/kg/peaksoft/peaksoftlmsab4/dto/authDto/AuthRequest.java
@@ -1,11 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.dto.authDto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AuthRequest {
-    private String email;
-    private String password;
-}
