+++ src/main/java/kg/peaksoft/peaksoftlmsab4/dto/exceptionDto/ExceptionResponse.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.http.HttpStatus;
 
+
 @Getter
 @Setter
 public class ExceptionResponse {
