+++ src/main/java/kg/peaksoft/peaksoftlmsab4/dto/mapper/view/StudentViewMapper.java
@@ -1,20 +1,20 @@
 package kg.peaksoft.peaksoftlmsab4.dto.mapper.view;
 
 import kg.peaksoft.peaksoftlmsab4.dto.response.StudentResponse;
-import kg.peaksoft.peaksoftlmsab4.entity.Student;
+import kg.peaksoft.peaksoftlmsab4.model.entity.Student;
 
 public class StudentViewMapper {
 
-    public StudentResponse viewStudent(Student student){
-        if(student==null){
+    public StudentResponse viewStudent(Student student) {
+        if (student == null) {
             return null;
         }
-        StudentResponse studentResponse=new StudentResponse();
+        StudentResponse studentResponse = new StudentResponse();
         studentResponse.setId(String.valueOf(student.getId()));
         studentResponse.setFirstName(student.getFirstName());
         studentResponse.setMobilePhone(student.getMobilePhone());
         studentResponse.setStudyFormat(student.getStudyFormat());
-        studentResponse.setEmail(student.getAuthInfo().getEmail());
+        studentResponse.setEmail(student.getEmail());
         return studentResponse;
     }
