+++ src/main/java/kg/peaksoft/peaksoftlmsab4/dto/response/AuthResponse.java
@@ -1,8 +1,10 @@
 package kg.peaksoft.peaksoftlmsab4.dto.response;
 
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+@Builder
 @Getter
 @Setter
 public class AuthResponse {
