--- src/main/java/kg/peaksoft/peaksoftlmsab4/entity/AuthInfo.java
@@ -1,41 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.entity;
-
-import kg.peaksoft.peaksoftlmsab4.enumPackage.Role;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import javax.persistence.*;
-import java.util.Collection;
-import java.util.Collections;
-
-@Entity
-@Table(name = "authInfo")
-@Getter
-@Setter
-
-public class AuthInfo implements UserDetails {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    Long id;
-    String email;
-    String password;
-    @Enumerated(EnumType.STRING)
-    private Role role;
-    private boolean isAccountNonExpired = true;
-    private boolean isAccountNonLocked = true;
-    private boolean isCredentialsNonExpired = true;
-    private boolean isEnabled = true;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singleton(role);
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-
-}
