--- src/main/java/kg/peaksoft/peaksoftlmsab4/model/enams/Authority.java
@@ -1,16 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.model.enams;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum
-
-Authority implements GrantedAuthority {
-    STUDENT,
-    INSTRUCTOR,
-    ADMIN;
-
-    @Override
-    public String getAuthority() {
-        return this.name();
-    }
-}
