+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/entity/AuthInfo.java
@@ -1,8 +1,7 @@
 package kg.peaksoft.peaksoftlmsab4.model.entity;
 
-import kg.peaksoft.peaksoftlmsab4.model.enams.Authority;
+import kg.peaksoft.peaksoftlmsab4.model.enums.Role;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -10,20 +9,19 @@ import org.springframework.security.core.userdetails.UserDetails;
 import javax.persistence.*;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.UUID;
 
 @Entity
-@Table(name = "authInfo")
+@Table(name = "auth_info")
 @Getter
 @Setter
-@NoArgsConstructor
 public class AuthInfo implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private UUID id;
-    private String email;
-    private String password;
-    private Authority authority;
+    Long id;
+    String email;
+    String password;
+    @Enumerated(EnumType.STRING)
+    private Role role;
     private boolean isAccountNonExpired = true;
     private boolean isAccountNonLocked = true;
     private boolean isCredentialsNonExpired = true;
@@ -31,11 +29,12 @@ public class AuthInfo implements UserDetails {
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
-        return Collections.singleton(authority);
+        return Collections.singleton(role);
     }
 
     @Override
     public String getUsername() {
         return email;
     }
+
 }
