+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/entity/Instructor.java
@@ -1,20 +1,15 @@
-package kg.peaksoft.peaksoftlmsab4.entity;
+package kg.peaksoft.peaksoftlmsab4.model.entity;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
 
 @Entity
-@Table(name = "admins")
+@Table(name = "instructors")
 @Getter
 @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-
-public class Admin {
+public class Instructor {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     Long id;
@@ -22,6 +17,10 @@ public class Admin {
     String firstName;
     @Column(name = "last_name")
     String lastName;
-    @OneToOne
+    @Column(name = "mobile_phone")
+    String mobilePhone;
+    String specialization;
+    @OneToOne(cascade = CascadeType.ALL)
     AuthInfo authInfo;
+
 }
