+++ src/main/java/kg/peaksoft/peaksoftlmsab4/model/enums/Role.java
@@ -1,11 +1,11 @@
-package kg.peaksoft.peaksoftlmsab4.enumPackage;
+package kg.peaksoft.peaksoftlmsab4.model.enums;
+
 
 import org.springframework.security.core.GrantedAuthority;
 
 public enum Role implements GrantedAuthority {
     ADMIN,
-    INSTRUCTOR,
-    STUDENT;
+    INSTRUCTOR;
 
     @Override
     public String getAuthority() {
