--- src/main/java/kg/peaksoft/peaksoftlmsab4/repositories/AuthInfoRepository.java
@@ -1,13 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.repositories;
-
-import kg.peaksoft.peaksoftlmsab4.model.entity.AuthInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface AuthInfoRepository extends JpaRepository<AuthInfo, UUID> {
-    Optional<AuthInfo> findByEmail(String email);
-}
