+++ src/main/java/kg/peaksoft/peaksoftlmsab4/repository/AdminRepository.java
@@ -1,6 +1,6 @@
 package kg.peaksoft.peaksoftlmsab4.repository;
 
-import kg.peaksoft.peaksoftlmsab4.entity.Admin;
+import kg.peaksoft.peaksoftlmsab4.model.entity.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
