+++ src/main/java/kg/peaksoft/peaksoftlmsab4/repository/AuthInfoRepository.java
@@ -1,12 +1,12 @@
 package kg.peaksoft.peaksoftlmsab4.repository;
 
-import kg.peaksoft.peaksoftlmsab4.entity.AuthInfo;
+import kg.peaksoft.peaksoftlmsab4.model.entity.AuthInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface AuthInfoRepository extends JpaRepository<AuthInfo,Long> {
+public interface AuthInfoRepository extends JpaRepository<AuthInfo, Long> {
     Optional<AuthInfo> findByEmail(String email);
 }
