+++ src/main/java/kg/peaksoft/peaksoftlmsab4/service/AdminService.java
@@ -1,5 +1,4 @@
 package kg.peaksoft.peaksoftlmsab4.service;
 
 public interface AdminService {
-    
 }
