+++ src/main/java/kg/peaksoft/peaksoftlmsab4/service/serviceImpl/AdminServiceImpl.java
@@ -1,4 +1,8 @@
 package kg.peaksoft.peaksoftlmsab4.service.serviceImpl;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class AdminServiceImpl {
+
 }
