+++ src/main/java/kg/peaksoft/peaksoftlmsab4/service/serviceImpl/AuthService.java
@@ -1,21 +1,19 @@
-package kg.peaksoft.peaksoftlmsab4.service;
+package kg.peaksoft.peaksoftlmsab4.service.serviceImpl;
 
 
 import kg.peaksoft.peaksoftlmsab4.config.jwt.JwtUtils;
-import kg.peaksoft.peaksoftlmsab4.dto.authDto.AuthRequest;
-import kg.peaksoft.peaksoftlmsab4.dto.authDto.AuthResponse;
+import kg.peaksoft.peaksoftlmsab4.dto.request.AuthRequest;
+import kg.peaksoft.peaksoftlmsab4.dto.response.AuthResponse;
 import lombok.AllArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
 @Service
 @AllArgsConstructor
 public class AuthService {
 
-    private final UserDetailsService userDetailsService;
     private final AuthenticationManager authenticationManager;
     private final JwtUtils jwtUtils;
