--- src/main/java/kg/peaksoft/peaksoftlmsab4/service/serviceImpl/AuthServiceImpl.java
@@ -1,5 +1,0 @@
-package kg.peaksoft.peaksoftlmsab4.service.serviceImpl;
-
-public class AuthServiceImpl {
-
-}
