+++ src/main/java/kg/peaksoft/peaksoftlmsab4/service/serviceImpl/StudentServiceImpl.java
@@ -1,4 +1,7 @@
 package kg.peaksoft.peaksoftlmsab4.service.serviceImpl;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class StudentServiceImpl {
 }
