+++ src/main/resources/application.properties
@@ -1,0 +1,12 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/peaksoftlms-a
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL10Dialect
+server.error.include-message=always
+
+#JWT
+application.jwt.secret-key=java4IsBestGroup
+application.jwt.token-prefix= Bearer
+application.jwt.expiration-date-after-days=30
